♻️ refactor: apply feedback
