+++ src/app/(with-header-sidebar)/dashboard/[id]/edit/_lib/memberService.ts
@@ -1,13 +1,15 @@
 import axiosInstance from '@/lib/axiosInstance';
 import { GetMembersResponse } from '@/types/member';
 
+const PAGE_SIZE = 4;
+
 export const getMembers = async (
   dashboardId: string,
   page: number = 1
 ): Promise<GetMembersResponse> => {
   try {
     const response = await axiosInstance.get(
-      `members?page=${page}&size=4&dashboardId=${dashboardId}`
+      `members?page=${page}&size=${PAGE_SIZE}&dashboardId=${dashboardId}`
     );
     return response.data;
   } catch (error) {
